requirements.txt
pandas~=1.2.3
sklearn~=0.0
scikit-learn~=0.24.1
numpy~=1.20.2
pillow~=8.2.0